Add separate component for daterange picker
